Migrate to DictionaryExportMenu, allow exporting keys from saved cards
